Add PR review fixes: correct PR number, extract MCPDetailsView, update docs
- Fix CHANGELOG and validation script PR number from #389 to #390
- Extract MCPDetailsView from AdminModal.jsx into its own component
- Update admin-panel.md with current polling intervals, new status fields
- Discard unrelated package-lock.json changes
