+++ docs/admin/admin-panel.md
@@ -1,6 +1,6 @@
 # Admin Panel
 
-Last updated: 2026-01-21
+Last updated: 2026-03-07
 
 The application includes an admin panel that provides access to configuration values, MCP server controls, application logs, and user feedback.
 
@@ -17,15 +17,15 @@ The application includes an admin panel that provides access to configuration va
 The admin dashboard automatically polls the backend to keep the MCP server status up to date. The polling implements exponential backoff to prevent overwhelming the backend when it's disconnected or experiencing issues.
 
 **Polling Behavior:**
-- **Normal polling**: Every 15 seconds when the backend is responsive
+- **Normal polling**: Every 30 seconds when the backend is responsive
 - **Failure handling**: When the backend is unreachable, polling backs off exponentially:
   - 1st retry: 1 second
   - 2nd retry: 2 seconds
   - 3rd retry: 4 seconds
   - 4th retry: 8 seconds
   - 5th retry: 16 seconds
-  - 6th+ retries: 30 seconds (max)
-- **Recovery**: When the backend becomes responsive again, polling returns to the normal 15-second interval
+  - 6th+ retries: 5 minutes (max)
+- **Recovery**: When the backend becomes responsive again, polling returns to the normal 30-second interval
 
 This prevents toast notification spam and reduces load on the backend during downtime or connection issues.
 
@@ -60,9 +60,12 @@ Returns the current MCP connection status and auto-reconnect settings.
 
 The response includes:
 
-- `connected_servers`: list of servers with active clients.
+- `connected_servers`: list of servers with active clients (excludes failed and non-configured servers).
 - `configured_servers`: all servers defined in `mcp.json`.
 - `failed_servers`: per-server error details and timing information (`attempt_count`, `error`, `last_attempt`, `backoff_delay`, `next_retry_in_seconds`).
+- `mcp_config_path`: absolute path to the active `mcp.json` file being read/written.
+- `tool_counts`: per-server count of discovered tools (only for configured servers).
+- `prompt_counts`: per-server count of discovered prompts (only for configured servers).
 - `auto_reconnect`: configuration and runtime state for the background reconnect task (`enabled`, `base_interval`, `max_interval`, `backoff_multiplier`, `running`).
 
 You can use this endpoint to quickly see which MCP servers are healthy, which are failing, and when the next reconnect attempt will occur.
